+++ myokit/_sim/fiber_tissue.py
@@ -20,9 +20,6 @@ class FiberTissueSimulation(myokit.CModule):
     """
     Runs a simulation of a fiber leading up to a rectangular piece of tissue.
 
-    Note: This simulation class is deprecated as of version 1.33.0, and will be
-    removed in future versions of Myokit.
-
     Takes the following input arguments:
 
     ``fiber_model``
@@ -131,10 +128,11 @@ class FiberTissueSimulation(myokit.CModule):
         super(FiberTissueSimulation, self).__init__()
 
         # Deprecated since 2021-05-28
-        import warnings
-        warnings.warn(
-            'The class `myokit.FiberTissueSimulation` is deprecated, and will'
-            ' be removed in future versions of Myokit.')
+        # Un-deprecated on 2022-04-20, following Adam's request
+        #import warnings
+        #warnings.warn(
+        #    'The class `myokit.FiberTissueSimulation` is deprecated, and will'
+        #    ' be removed in future versions of Myokit.')
 
         # List of globally logged inputs
         self._global = ['time', 'pace']